requirements.txt
fastapi
uvicorn
python-dotenv
easyocr
google-genai
python-multipart
